remove populate file and move functionality to db file
